Check for column existence before removing it

+++ db/migrate/20191028204156_drop_indexed_sort_field_column_from_metadata_profile_elements.rb
@@ -1,5 +1,7 @@
 class DropIndexedSortFieldColumnFromMetadataProfileElements < ActiveRecord::Migration[5.2]
   def change
-    remove_column :metadata_profile_elements, :indexed_sort_field
+    if column_exists? :metadata_profile_elements, :indexed_sort_field
+      remove_column :metadata_profile_elements, :indexed_sort_field
+    end
   end
 end
